Remove rails dependency from gemspec
We do not want rails as a gem dependency because a bunch of other gems are
dependent on rails, such as activerecord. Obviously if we are using Sequel in a
Rails app, there is no need for ActiveRecord. We do not want to waste cycles
loading code that we do not need.